requirements.txt
openai~=1.3.9
numpy~=1.26.2
matplotlib~=3.8.2
scipy~=1.11.4
SpeechRecognition
Pyaudio
sounddevice~=0.4.6
openai-whisper
torch~=2.1.2+cu121
setuptools-rust

python-dotenv~=1.0.0
torchaudio~=2.1.2+cu121
gTTS
playsound